refactor: extract shared VRL runtime crate
