+++ src/core/src/service/search/super_cluster/leader.rs
@@ -45,14 +45,13 @@ use tracing_opentelemetry::OpenTelemetrySpanExt;
 use crate::service::search::{
     SEARCH_SERVER, SearchResult,
     datafusion::{
-        context::{QueryExecutionContext, SearchContextBuilder, register_table},
+        context::{SearchContextBuilder, register_table},
         optimizer::{
             context::{PhysicalOptimizerContext, RemoteScanContext, StreamingAggregationContext},
             create_physical_plan,
         },
     },
     inspector::{SearchInspectorFieldsBuilder, search_inspector_fields},
-    prepare_query_transforms,
     sql::Sql,
     utils::{AsyncDefer, ScanStatsVisitor, check_query_default_limit_exceeded},
 };
@@ -233,7 +232,6 @@ async fn run_datafusion(
 
     // construct physical plan
     let is_complete_cache_hit = Arc::new(Mutex::new(false));
-    let query_context = QueryExecutionContext::new(prepare_query_transforms(&req.org_id));
     let ctx = SearchContextBuilder::new()
         .target_partitions(cfg.limit.cpu_num)
         .add_context(PhysicalOptimizerContext::RemoteScan(RemoteScanContext {
@@ -246,7 +244,7 @@ async fn run_datafusion(
             StreamingAggregationContext::new(&req, is_complete_cache_hit.clone()).await?,
         ))
         .add_context(PhysicalOptimizerContext::AggregateTopk)
-        .build(&req, &sql, &query_context)
+        .build(&req, &sql)
         .await?;
 
     // register table
