+++ src/o2_vrl/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "o2_vrl"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+publish = false
+
+[dependencies]
+async-trait.workspace = true
+config.workspace = true
+vector-enrichment.workspace = true
+vrl.workspace = true
+
+[dev-dependencies]
+serde_json.workspace = true
+tokio.workspace = true
