+++ src/search/Cargo.toml
@@ -12,21 +12,19 @@ vectorscan = []
 vortex = []
 
 [dependencies]
-async-trait.workspace = true
 chrono.workspace = true
 config.workspace = true
 datafusion.workspace = true
 hashbrown = { workspace = true, optional = true }
 infra.workspace = true
 log.workspace = true
+o2_vrl.workspace = true
 o2_enterprise = { workspace = true, optional = true, default-features = false }
 parking_lot = { workspace = true, optional = true }
 serde.workspace = true
 tokio.workspace = true
 tracing.workspace = true
 utoipa.workspace = true
-vector-enrichment.workspace = true
-vrl.workspace = true
 
 [dev-dependencies]
 arrow.workspace = true
